Reorganize non-reusable components
