+++ gatsby-browser.ts
@@ -9,5 +9,5 @@ import { pdfjs } from "react-pdf";
 // https://github.com/wojtekmaj/react-pdf/issues/321#issuecomment-451291757
 pdfjs.GlobalWorkerOptions.workerSrc = `//cdnjs.cloudflare.com/ajax/libs/pdf.js/${pdfjs.version}/pdf.worker.js`;
 
-import customWrapPageElement from "./src/components/WrapPageElements";
+import customWrapPageElement from "./src/layouts/WrapPageElements";
 export const wrapPageElement = customWrapPageElement;
