+++ gatsby-ssr.ts
@@ -1,6 +1,6 @@
-import customWrapPageElement from "./src/components/WrapPageElements"
-export const wrapPageElement = customWrapPageElement
+import customWrapPageElement from "./src/layouts/WrapPageElements";
+export const wrapPageElement = customWrapPageElement;
 
 export const onRenderBody = ({ setHtmlAttributes }) => {
-  setHtmlAttributes({ lang: "en" })
-}
+  setHtmlAttributes({ lang: "en" });
+};
