+++ src/layouts/Layout.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 
-import Nav from "./Nav"
-import Footer from "./Footer"
+import Nav from "../components/Nav"
+import Footer from "../components/Footer"
 
 interface Props {
   children: React.ReactNode
